Refactory mouse fallback.
